Better labelling and adding another table config
